+++ lib/src/material/app.dart
@@ -2,6 +2,8 @@ import 'package:dart_eval/dart_eval.dart';
 import 'package:dart_eval/dart_eval_bridge.dart';
 import 'package:dart_eval/stdlib/core.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_eval/src/foundation/key.dart';
+import 'package:flutter_eval/src/material/theme_data.dart';
 import 'package:flutter_eval/src/sky_engine/ui/painting.dart';
 import 'package:flutter_eval/src/widgets/app.dart';
 import 'package:flutter_eval/src/widgets/framework.dart';
@@ -12,10 +14,28 @@ class $MaterialApp implements $Instance {
   static const $declaration = BridgeClassDef(BridgeClassType($type, isAbstract: false, $extends: $WidgetsApp.$type),
       constructors: {
         '': BridgeConstructorDef(BridgeFunctionDef(returns: BridgeTypeAnnotation($type), namedParams: [
-          BridgeParameter('color', BridgeTypeAnnotation($Color.$type), true),
+          BridgeParameter('key', BridgeTypeAnnotation($Key.$type), true),
+          BridgeParameter('navigatorKey', BridgeTypeAnnotation($Key.$type), true),
           BridgeParameter('home', BridgeTypeAnnotation($Widget.$type), true),
-          BridgeParameter('title', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.stringType)), true),
+          BridgeParameter('routes', BridgeTypeAnnotation(BridgeTypeRef.spec(BridgeTypeSpec('dart:core', 'Map'))), true),
+          BridgeParameter('initialRoute', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.stringType)), true),
+          BridgeParameter('onUnknownRoute', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.functionType)), true),
           BridgeParameter('builder', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.functionType)), true),
+          BridgeParameter('title', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.stringType)), true),
+          BridgeParameter('onGenerateTitle', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.functionType)), true),
+          BridgeParameter('color', BridgeTypeAnnotation($Color.$type), true),
+          BridgeParameter('theme', BridgeTypeAnnotation($ThemeData.$type), true),
+          BridgeParameter('darkTheme', BridgeTypeAnnotation($ThemeData.$type), true),
+          BridgeParameter(
+              'debugShowMaterialGrid', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.boolType)), true),
+          BridgeParameter(
+              'showPerformanceOverlay', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.boolType)), true),
+          BridgeParameter(
+              'checkerboardRasterCacheImages', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.boolType)), true),
+          BridgeParameter(
+              'checkerboardOffscreenLayers', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.boolType)), true),
+          BridgeParameter(
+              'debugShowCheckedModeBanner', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.boolType)), true)
         ]))
       },
       methods: {},
@@ -26,15 +46,36 @@ class $MaterialApp implements $Instance {
 
   static $Value? $new(Runtime runtime, $Value? target, List<$Value?> args) {
     return $MaterialApp.wrap(MaterialApp(
-      color: args[0]?.$value,
-      home: args[1]?.$value,
-      title: args[2]?.$value ?? '',
-      builder: args[3] == null
+      key: args[0]?.$value,
+      navigatorKey: args[1]?.$value,
+      home: args[2]?.$value,
+      routes: args[3] == null ? const {} : $reifyRoutes(runtime, args[3] as $Map),
+      initialRoute: args[4]?.$value,
+      onUnknownRoute: args[5] == null
           ? null
-          : (ctx, widget) => (args[3] as EvalCallable)(
-                  runtime, null, [$BuildContext.wrap(ctx), widget == null ? const $null() : $Widget.wrap(widget)])
-              as Widget,
+          : (settings) => (args[5]! as EvalCallable).call(runtime, null, [$Object(settings)])?.$value,
+      builder: args[6] == null
+          ? null
+          : (context, child) => (args[6]! as EvalCallable).call(runtime, null,
+              [$BuildContext.wrap(context), child == null ? const $null() : $Widget.wrap(child)])?.$value,
+      title: args[7]?.$value,
+      onGenerateTitle: args[8] == null
+          ? null
+          : (context) => (args[8]! as EvalCallable).call(runtime, null, [$BuildContext.wrap(context)])?.$value,
+      color: args[9]?.$value,
+      theme: args[10]?.$value,
+      darkTheme: args[11]?.$value,
+      debugShowMaterialGrid: args[12]?.$value ?? false,
+      showPerformanceOverlay: args[13]?.$value ?? false,
+      checkerboardRasterCacheImages: args[14]?.$value ?? false,
+      checkerboardOffscreenLayers: args[15]?.$value ?? false,
+      debugShowCheckedModeBanner: args[16]?.$value ?? true,
     ));
+  }
+
+  static Map<String, Widget Function(BuildContext)> $reifyRoutes(Runtime runtime, $Map routes) {
+    return routes.$value.map((key, value) => MapEntry(
+        key.$value, (context) => (value as EvalCallable).call(runtime, null, [$BuildContext.wrap(context)])!.$value));
   }
 
   $MaterialApp.wrap(this.$value);